Update the wearable when updating boardedDeparture
This uses a GoogleApiClient instance variable that is co-opted to mean "if mGac
is not null, then this is the boarded departure".  It updates the wearable
whenever mergeEstimate() is called.

The Wear API update moves to the Departure object, so it can get called by the
BART API network handler threads.